+++ TopologyParser/TopologyParser.cpp
@@ -15,12 +15,12 @@ NeuralNetwork::Topology TopologyParser::parse(nlohmann::json &j) {
         Phenotype::point output = {it["output"][0], it["output"][1]};
         if (output[0] > max_layers)
             max_layers = output[0];
-        float const input_weight = it["input_weight"];
-        float const memory_weight = it["memory_weight"];
-        float const reset_input_weight = it["reset_input_weight"];
-        float const update_input_weight = it["update_input_weight"];
-        float const reset_memory_weight = it["reset_memory_weight"];
-        float const update_memory_weight = it["update_memory_weight"];
+        double const input_weight = it["input_weight"];
+        double const memory_weight = it["memory_weight"];
+        double const reset_input_weight = it["reset_input_weight"];
+        double const update_input_weight = it["update_input_weight"];
+        double const reset_memory_weight = it["reset_memory_weight"];
+        double const update_memory_weight = it["update_memory_weight"];
         bool disabled = it["disabled"];
         new_phenotypes.push_back(
                 new Phenotype(input, output, input_weight, memory_weight, reset_input_weight, update_input_weight,
